requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
pydantic[email]>=2.8.0
aiofiles>=23.2.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
requests>=2.31.0
